requirements.txt
astroid==2.0.4
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.2
decorator==4.3.0
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.0
mccabe==0.6.1
networkx==2.2
numpy==1.15.2
opencv-python==3.4.3.18
Pillow==5.2.0
pkg-resources==0.0.0
pyparsing==2.2.2
pytesseract==0.2.4
python-dateutil==2.7.3
PyWavelets==1.0.1
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
toolz==0.9.0
typed-ast==1.1.0
wrapt==1.10.11
